modify python package
